shared: refactor move files to shared
