--- src/app/shared/common/entry.interface.ts
@@ -1,11 +1,0 @@
-/**
- * Created by vicensfayos on 15/08/16.
- */
-
-export interface Hero {
-    uuid: string;
-    data: any;
-    sleepingQuality: string;
-    tirednessFeeling: string;
-
-}
