+++ src/app/shared/common/entry.ts
